Make it possible to pass an "embed" prop to Resource
The plan is to have resources retrieved from the backend take a format
like this:

{
  links: [
    {
      href: 'https://api.com/resources/1',
      rel: 'thing',
      embed: {
        // The response you would receive from GETting this URI
      }
    }
  ],
  // other data
}

This way, we can give pre-loaded responses to queries that commonly come
after the current query. The pre-loaded response can be passed to
Resource through the "embed" prop. Resource immediately calls
this.loadResource on whatever was passed into its embed prop, if
anything.